+++ lib/public/view.js
@@ -15,7 +15,7 @@ import { h, switchCase } from '/js/src/index.js';
 import NavBar from './components/NavBar/index.js';
 import LogsOverview from './views/Logs/Overview/index.js';
 import { LogCreationPage } from './views/Logs/Create/LogCreationPage.js';
-import { RunsOverviewPage } from './views/Runs/Overview/index.js';
+import { RunsOverviewPage } from './views/Runs/Overview/RunsOverviewPage.js';
 import { RunDetailsPage } from './views/Runs/Details/RunDetailsPage.js';
 import SubsystemOverview from './views/Subsystems/Overview/index.js';
 import SubsystemDetail from './views/Subsystems/Details/index.js';
